Provide per-mapping setting of stay_in_mode
When using nested modes with "stay_in_mode", current versions of awesome
will throw an error, because two keygrabbers are running at the same
time. As a workaround, it is now possible to set `stay_in_mode` per
key in a mapping table. The global option is still around, but the local
setting takes precedence when a key where it is present is pressed.